fmt + accommodate the new test elsewhere
