+++ hil-test/Cargo.toml
@@ -12,6 +12,10 @@ name              = "alloc_psram"
 harness           = false
 
 [[bin]]
+name              = "bootloader"
+harness           = false
+
+[[bin]]
 name              = "crypto"
 harness           = false
